dom/tests/mochitest/webapps/test_bug_771294.xul

<?xml version="1.0"?>

<!-- Any copyright is dedicated to the Public Domain.
   - http://creativecommons.org/publicdomain/zero/1.0/ -->

<?xml-stylesheet type="text/css" href="chrome://global/skin"?>
<?xml-stylesheet type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"?>

<window xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        title="Mozilla Bug 771294">
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>
  <script type="application/javascript" src="apphelper.js"/>
  <script type="application/javascript" src="jshelper.js"/>

<script>
SimpleTest.waitForExplicitFinish();

Components.utils.import('resource://gre/modules/Services.jsm');

var blocked = true;

Services.obs.addObserver(
  function observeShowing() {
    Services.obs.removeObserver(observeShowing, "PopupNotifications-showing");
    blocked = false;
  },
  "PopupNotifications-showing",
  false
);

Services.obs.addObserver(
  function observeInstalling() {
    Services.obs.removeObserver(observeInstalling, "webapps-ask-install");
    // Spin the event loop before running the test to give the registry time
    // to process the install request and (hopefully not) show the doorhanger.
    setTimeout(function verify() {
      ok(blocked, "Install panel was blocked after immediate redirect");
      tearDown();
      SimpleTest.finish();
    }, 0);
  },
  "webapps-ask-install",
  false
);

</script>

  <!-- Load a page that initiates an app installation and then immediately
     - redirects to a page at a different origin.  We can't do this directly
     - inside this test page, because that would cause the test to hang. -->
  <iframe src="http://www.example.com:80/chrome/dom/tests/mochitest/webapps/apps/bug_771294.xul"/>

</window>
